[fix] rain/snow attribute type from long to int

+++ Hwaxby_back/src/main/java/spring/Hwaxby_back/domain/OpenWeather/CurrentWeather.java
@@ -104,7 +104,7 @@ public class CurrentWeather extends OpenWeather{
 
         /** 지난 1 시간 동안의 강우량, mm */
         @JsonProperty("1h")
-        private long rain1h;
+        private int rain1h;
     }
 
     @Data
@@ -112,7 +112,7 @@ public class CurrentWeather extends OpenWeather{
 
         /** 지난 1 시간 동안의 눈량, mm */
         @JsonProperty("1h")
-        private long snow1h;
+        private int snow1h;
     }
 
 }
